chore(ts): Move to typescript
