+++ package.json
@@ -4,6 +4,7 @@
   "description": "Robust Environment Configuration for Universal Applications.",
   "main": "lib/index.cjs.js",
   "module": "lib/index.esm.js",
+  "types": "lib/index.d.ts",
   "bin": {
     "unienv": "bin/unienv"
   },
@@ -55,6 +56,7 @@
   "devDependencies": {
     "@babel/core": "^7.5.5",
     "@effective/eslint-config": "^0.7.12",
+    "@types/yargs": "^13.0.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^24.8.0",
     "babel-preset-edge": "^5.3.0",
