+++ src/noenv.test.js
@@ -9,7 +9,7 @@ const snapshotOpts = {
 
 /* eslint-disable import/no-commonjs */
 // We can't use ESM when relying on the fact the the env from the top is correctly respected.
-const api = require(".")
+const api = require("..")
 
 test("Defaults to development for NODE_ENV", () => {
   const { raw, stringified, webpack } = api.getEnvironment()
